Bug 1402991 - Add docs for landing patches to git and mozilla-inbound
Summary:
Adds documentation to guide users when landing a revision to either
a git repository or to firefox, i.e. mozilla-inbound.

This also adds a small section to clarify the behavior of
`arc patch` which is not immediately clear unless one pays attention.
The main thing is that arc patch, by default, doesn't apply changes
to the current working set, which sort of goes against what the word
"patch" means - the behavior is fine, this doc is just to clear up
confusion.

Test Plan: Read through the changes and perhaps view it in an .rst viewer to make sure formatting is ok.

Reviewers: mcote

Reviewed By: mcote

Bug #: 1402991

Differential Revision: https://phabricator.services.mozilla.com/D95