Let moz-phab install arcanist for us
When I followed the instructions I hit a slight problem (because I
thought I had my path set but didn't, due to unrelated reasons).  I
found that moz-phab will download and install arcanist for me.  When
doing this it puts it in ~/.mozbuild/moz-phab/arcanist and not
~/.mozbuild/arcanist  I propose to update these install instructions to
reflect that path and also let moz-phab install arcanist.