Annotate some code-blocks
In particular, the `export PATH` command is just shown as normal text right now on https://moz-conduit.readthedocs.io/en/latest/mozphab-linux.html .

To fix this, I'm cargo-culting this code-block annotation from elsewhere in this .rst file (and adding an initial "$" for consistency), and I'm hoping it works to fix the formatting of this code.  (I'm adding the annotation for the final code-block, too, even though it's already displayed correctly. I'm assuming it's still good to annotate it for consistency/robustness.)