Reorder steps
Reorder steps so programs are installed prior to adding them to path. Installing git and moz-phab are moved before php since they're less involved and they needed for updating path step. The php -i sanity check is left after the path is updated to ensure php is on the path when running the command.